Large cleanup of maq and various improvements
This is another early big check-in that attempts to unify the query
code for reviews and bands as well as fix a number of minor nits.

I've also allowed the throttle option to be a float in order to make
the default throttle a half a second which works fine in my testing
(a balance between performance and politeness).

Put progress bar under the verbose operating mode as the command line
should be silent regarding debug messages unless told otherwise.

NOTE: I still can not get 'iDisplayStart' to work yet so you can skip
batches and start in the middle of MA's JSON list. Work in progress.