Add batch and band logo support
Be able to download band logo files on request.

Band logos are downloaded and named 'band id'.'jpg|gif' etc.

Add '--batch' parameter to allow fetching records in the middle
of a query. Reviews come in batches of 200 while bands come in batches
of 500. It seems statically set with no way to discover these sizes.

Minor touch up on verbose output and managing the progress bar.