Add quiet mode and make verbose mode standard
I realize that the majority of users will probably want verbose mode
to be the default since it is the most useful (particularly if you
are scraping a lot of bands/reviews off of MA).

But for those who want to run this out of a cron, add a quiet
mode to preserve the ability for quiet operation.